Migrate student-verify page
